+++ hi-ml-azure/testazure/testazure/test_datasets.py
@@ -5,15 +5,18 @@
 """
 Test the data input and output functionality
 """
+from pathlib import Path
 from unittest import mock
-from azureml.exceptions._azureml_exception import UserErrorException
+from health_azure.utils import PathOrString
 
 import pytest
+from azureml._restclient.exceptions import ServiceException
 from azureml.core import Dataset
 from azureml.data import FileDataset, OutputFileDatasetConfig
 from azureml.data.azure_storage_datastore import AzureBlobDatastore
 from azureml.data.dataset_consumption_config import DatasetConsumptionConfig
-from azureml._restclient.exceptions import ServiceException
+from azureml.exceptions._azureml_exception import UserErrorException
+
 from health_azure.datasets import (DatasetConfig, _input_dataset_key, _output_dataset_key,
                                    _replace_string_datasets, get_datastore, get_or_create_dataset)
 from testazure.util import DEFAULT_DATASTORE, DEFAULT_WORKSPACE
@@ -79,6 +82,36 @@ def test_dataset_input() -> None:
     aml_dataset = dataset_config.to_input_dataset(workspace=workspace, dataset_index=1)
     assert isinstance(aml_dataset, DatasetConsumptionConfig)
     assert aml_dataset.mode == "mount"
+
+
+@pytest.mark.parametrize("target_folder", [
+    "",
+    None,
+])
+def test_dataset_input_target_empty(target_folder: PathOrString) -> None:
+    """
+    Leaving the target folder empty should NOT create a path_on_compute that is "."
+    """
+    workspace = DEFAULT_WORKSPACE.workspace
+    # This dataset must exist in the workspace already, or at least in blob storage.
+    dataset_config = DatasetConfig(name="hello_world", datastore=DEFAULT_DATASTORE, target_folder=target_folder)
+    aml_dataset = dataset_config.to_input_dataset(workspace=workspace, dataset_index=1)
+    assert isinstance(aml_dataset, DatasetConsumptionConfig)
+    assert aml_dataset.path_on_compute is None
+
+
+@pytest.mark.parametrize("target_folder", [
+    ".",
+    Path(),
+    Path("."),
+])
+def test_dataset_invalid_target(target_folder: PathOrString) -> None:
+    """
+    Passing in "." as a target_folder shouold raise an exception.
+    """
+    with pytest.raises(ValueError) as ex:
+        DatasetConfig(name="hello_world", datastore=DEFAULT_DATASTORE, target_folder=target_folder)
+    assert "current working directory" in str(ex)
 
 
 def test_dataset_output() -> None: